requirements.txt
pyodbc
flask
flask-restful
tenacity
python-dotenv
PyJWT